improve styling
