+++ neural_insights/gui/src/components/Diagnosis/Diagnosis.scss
@@ -48,6 +48,13 @@
     margin-bottom: 6px;
   }
 
+  .data-panel-top {
+    background-color: white;
+    padding: 10px;
+    border-radius: 0.5rem;
+    margin-top: 6px;
+  }
+
   .flexbox {
     display: flex;
     max-height: 99vh;
@@ -60,7 +67,14 @@
     display: flex;
     flex-direction: column;
     max-height: 90vh;
-    width: 300px;
+    width: 350px;
+  }
+
+  .workloads-flex {
+    flex: 1;
+    display: flex;
+    overflow: auto;
+    width: 350px;
   }
 
   .overflow-table {
@@ -73,7 +87,7 @@
   .flex-item {
     padding: 3px;
     width: 98%;
-    max-width: calc(100vw - 320px);
+    max-width: calc(100vw - 370px);
   }
 
   .accuracy-table {
